core: move Blockchainer interface to a separate package
